chore: more making the diff useful
