+++ packages/snap/src/ui.ts
@@ -1,8 +1,7 @@
 import { copyable, divider, heading, panel, text } from '@metamask/snaps-sdk';
 import { invalidParameters } from './errors';
-import { prettyPrint } from './transaction-ui/commands';
-import { transactionTitle } from './transaction-ui/transaction-title';
-import type { getFormatNumber } from './transaction-ui/utils';
+import { transactionTitle } from './transaction-title';
+import type { getFormatNumber } from './utils';
 import {
   formatAssetAmount,
   formatMarketCode,
@@ -19,7 +18,7 @@ import {
   indentText,
   isUnspecified,
   minimiseId,
-} from './transaction-ui/utils';
+} from './utils';
 import type { VegaTransaction, EnrichmentData } from './types';
 
 /**
@@ -956,4 +955,31 @@ export function prettyPrintStopOrdersCancellation(
     ];
   }
   return [textFn(`Cancel all stop orders from all markets`)];
+}
+
+/**
+ * Recurively pretty prints an object as snap-ui elements.
+ *
+ * @param obj - Object to pretty print. Primitives will be coerced to strings, while objects will be recursed into.
+ * @returns List of snap-ui elements.
+ */
+export function prettyPrint(obj: any) {
+  const elms = [];
+
+  if (obj === null || obj === undefined) {
+    return [text(`**Empty transaction provided**`)];
+  }
+
+  for (const [key, val] of Object.entries(obj)) {
+    if (val === null || val === undefined) {
+      elms.push(text(`**${key}**: empty value`));
+    } else if (typeof val === 'object') {
+      elms.push(text(`**${key}**: `));
+      elms.push(...prettyPrint(val));
+    } else {
+      elms.push(text(`**${key}**: ${val}`));
+    }
+  }
+
+  return elms;
 }
